feat: add getUserById functionality

+++ internal/user/repository.go
@@ -9,7 +9,7 @@ import (
 
 type Repository interface {
 	GetUsers(ctx context.Context) ([]*entity.User, error)
-	GetUserById(ctx context.Context, id int) (*entity.User, error)
+	GetUserById(ctx context.Context, id int64) (*entity.User, error)
 	GetUserByEmail(ctx context.Context, email string) (*entity.User, error)
 	CreateUser(ctx context.Context, user *dtos.CreateUserDto) (*entity.User, error)
 	UpdateUser(ctx context.Context, id int, user *dtos.UpdateUserDto) (*entity.User, error)
